Разбил вспомогательные методы на отдельные файлы
