+++ src/utils/common.js
@@ -1,0 +1,6 @@
+export const getRandomInteger = (from = 0, to = 1) => {
+  const lower = Math.ceil(Math.min(from, to));
+  const upper = Math.floor(Math.max(from, to));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
