+++ src/view/task.js
@@ -1,4 +1,4 @@
-import {isTaskExpired, isTaskRepeating, localizeDueDate} from "../utils.js";
+import {isTaskExpired, isTaskRepeating, localizeDueDate} from "../utils/task.js";
 import Abstract from "./abstract.js";
 
 const createTaskTemplate = (task) => {
